python/requirements.txt
python-dotenv
requests
websockets
asyncio
aiohttp
solana
